[FIX] web, website: ensure that portal content is always readable
Commit [1] changed the card style of Bootstrap cards. Only coloring
the header and footer while the body always has a white background.
With website installed, this can cause issue for users who choose to
have a dark background on their website.

Steps to reproduce:
- Install website and Sales
- Change the background of a website to a dark color
- Go to the portal and choose a sales order
- The content in the table is white and therefore not readable

Since we set the color of cards white, content inside of them should
have a dark color. However if the website has a text color that is light
the content of a table will reflect that color. Therefore, making sure
that the table has a specific color and is always readable is needed.

[1]: https://github.com/odoo/odoo/commit/ddf2d2dd666c6e3fa2b6b7f943575657a42d6d8b

task-2728923