[FIX] web_editor,website_forum,web_unsplash: portal image in forum post
A portal user cannot insert an image to a forum post

Steps to reproduce:
1. Install the Forum app
2. Connect as portal
3. Go to the forum and create a new post
4. Type '/image' and try to insert an image
5. An Access Error is raised preventing the portal user from inserting
an image

Solution:
Hide the /image and /link commands if it is specified in options.plugins
Add a route to fetch attachments. If the user is a portal, there are no
attachments available for him so we return an empty list.
Call create and search_read with sudo to allow access to portal users
if he has the rights.
Provide the res_model when creating attachment to check access.
Allow any user to use unsplash in the image widget and hide the unsplash
configurator for users with no unsplash rights.
Also fixed the condition that checks if the user has enough karma.
In 15.0, [this PR] removes the `/image` command for portal users.

[this PR]: https://github.com/odoo/odoo/pull/89250

opw-2648770
task-2811325