[FIX] website_sale: prevent edition of uneditable content
Commit [1] introduced template with the `stretched-link` class which
made some templates for the Product Snippet editable when they
shouldn't.

Steps to reproduce:
- Drop a product snippets
- Choose a template with a `stretched-link` bootstrap class
    (e.g) Horizontal Cards
- Click on the snippet

The editor will target the picture / link instead of the snippet.
This is caused by the fact that the `stretched-link` class changes
the way event-pointers are handled.

task-2677203

[1]: https://github.com/odoo/odoo/commit/49bd79b0bdca76415780bdfa199195371c2692e5