[IMP] tests: enables touch feature detection
This commit brings the ability to use some feature detection
mecanism to the test suite. The jquery.touchSwipe library
implementation requires the browser to get ontouch* events from
the window to be available to assert that the browser has touch.

The switch will allow that mecanism to work as expected, and
allow some tests using touch events to be correctly executed.