[FIX] website: fix scroll to anchor on same page
As URLs may or may not have a trailing slash (either way works). In the
anchor click code, it was not correct to compare the pathname of the
current page to the pathname of the target of a link without take it
into account.

Steps to reproduce the bug (e.g.):

- On the home page, create a link to an anchor with "/#section".
- Go to a page in french by adding "/fr" in the url.
- click on the link to the anchor. The page does a reload rather than a
scroll animation to the anchor. (because "/fr/" is not equal to "/fr").

After this commit, we remove the slashes to do the comparison. And so,
in the example above, "fr" is equal to "fr" as we want.

task-2172312