[IMP] mail: try to better update mail state when having sending failures
Purpose of this commit is to try to update MailMail ``failure_reason``,
``failure_type`` when possible. In order to ease update and maintenance
mail status update is partially moved in ``_postprocess_sent_message``.
When having a failure it may now update MailMail state information along
with notifications information. That way state update is deletated to the
post-process instead of having it done sometimes manually even when calling
the post processing.

Test update show that now mail and notifications are better aligned.

Task-2684479 (Mail: Better send error storage and display)