[MOV] mail, mass_mailing: extract mail values preparation into a method
When sending a ``MailMail`` some data preparation is done. This is done
directly in ``_send`` and in sub-methods. As a given MailMail may lead to
several emails being sent we have to go from a mail to a list of emails
to sent.

In this commit we move code so that all preparation is done in a sub-method
``_send_prepare_outgoing_list``. That way it can be cleanly overridden to
add or modify values before sending the actual emails.

This commit does not change behavior and calls done to ``build_email``. This
will be updated in the next commits, notably to improve "cc" management.

Task-2684479 (Mail: Better send error storage and display)