[IMP] base, mail, mass_mailing: better report from errors when sending emails
PURPOSE

Better report errors happening when sending emails.

SPECIFICATIONS

When having a global exception when sending emails (aka not a MemoryError and
not a psycopg or SMTP error), try to find issues related to

  * invalid email from;
  * invalid smtp from;
  * encoding issue in from or to;

And propagate this value to the mail.mail, and its sub records notifications
and mailing traces.

Task-2684479 (Mail: Better send error storage and display)