[FIX] website, web_editor: block options panel until changes are applied
Before this commit, modifying an option in the options tab did only put
a blocking loader overlay on the page, but not on the options tab.
This made it possible to modify several options while a single change
was being applied. This is a problem when using the gradient editor
because it shows the updated state only after it has been applied.

After this commit when a change is done in one of the colorpicker of the
options tab, an additional blocking loader overlay is put on top of the
options tab until de change is fully applied.

task-2633169