[IMP] website, web_editor: propose a new image comparison block
This commit introduces a brand new block that allows the user to put two
images on top of each other with a slider that allows to see the
differences between the two images.

Additional information:
It is possible to have descriptions for each of the pictures.
A button to enlarge the two images is optionally available.

task-2531109