[FIX] web_editor: unlink only the selected link
Before this commit when a button containing a text then an icon was
unlinked in Chrome, all further buttons were also unlinked.

After this commit an extra text node is added inside the selection to
prevent Chrome from wrongly determining the involved nodes.
Also, Firefox tries to keep the `class` of the `<a>` by replacing it
by a `<span>` during `unlink` - to prevent this, this commit also
removes the classes before calling the `unlink` command.

Steps to reproduce:
- drop Text+Image snippet twice
- inside first button insert a pictogram after the text
- inside first button insert a pictogram before or after the text
- click on the text of the first button => the link popover appears
- click on the unlink button of the popover
=> both buttons become flat texts in Chrome
=> nothing happens visually in Firefox, but the `<a>` actually gets
turned into a `<span>`

task-2778912