[REF] point_of_sale,*pos*: use RPC and ORM services
This commit is the start of the POS addons' shift to the
new env. This is accomplished by using the default env
introduced by the web addon in the `owl.Component`.

Based on this new env, we create `pos_env` which points
to other objects that are essential in the pos app.

Now that we are using the new env, we can now start
consuming the services offered by the web addon and
in this commit, we do so by changing our use of the
old `rpc.query` to the use of ORM and RPC services.

However, there are still parts of the code where using the
old env is unavoidable like the `iot_longpolling` which
depends on the `BusService` which is also a legacy service,
so we keep the reference to the old env as `legacyEnv` in
the singleton of `PosGlobalState`.