[IMP] website: keep table of contents title and menu entries in sync
The table of contents menu entries are generated automatically which
poses a problem in translation mode. The menu translation entries
should not be editable separately.

This commit prevents the menu entries from being translatable, and
shows a tooltip instead, explaining to the user that they are
generated from the title entries. Additionally, the translation widget
now checks for translation entries on the same page that have the same
translation id. If so, these are updated together, making sure that
they stay synchronised.

task-2752391