[FIX] survey: survey description spacing issue in print view
before when we print survey, the spacing (div) used to come even if there  is no
content in the field. as it is an HTML field it by default stores <p> tag having
a <br /> tag inside. this caused to fail the t-if condition. I have solved it
with is_html_empty() function.