[FIX] website: make `theme.website.menu` working
Before this commit, there were multiple issues related to the creation
of `theme.website.menu` records in a theme.

Indeed, the point of creating a menu is to add in to the existing
website's navbar, whose records are `website.menu`.
To do so, one should specify it by refering to the generic menu from
the XML file data.

The system then needs to understand that and find the corresponding
record for the given website. Note that there is no way to identify a
generic menu's website copies. Neither through a direct DB field,
neither through a persistent info like we do with the `key` attribute
of views.

Before this commit, when one would try to create a `theme.website.menu`
that menu would be created for every website when instaling the theme
on a particular website, which is terribly wrong as it goes against the
multi-website holy grail rule: when editing a website, do not impact
other websites.

Also, it was impossible to create a `theme.website.menu` having a
`website.menu` record as parent, which seems to be the main use case.