[IMP] ir.ui.view: remove `_get_inheriting_views_domain`
This has as effect to change the query sql to get inheriting
views when website is installed, even for back-end views,
which seems a bit counter-productive: The query is more complicated
and potentially slower for back-end views once website is installed.

Besides, this will open the door to easly ormcache this method
as it will depends on less fields.