[FIX] point_of_sale: disappearing text at Opening/Closing popup
Before this commit, the text from the "Notes"
area would disappear on changing the amount
at the input area "Opening"/"Counted" cash of
both the CashOpening and ClosePos popups.

This behaviour could result in the user having to
insert a long text input two times making for an
unpleasant UX.

In this commit we address this issue by making sure
that anything that the user inserts at the text box
is being erased. Instead the new information coming
from the money details popup is added at the end of
the text.

task-2755384