[FIX] website_blog: refresh options UI after adding a new tag
Since [1] adding a new tag on a blog does not refresh the tag list in
the options tab. That commit sorts the `methodsNames` aplhabetically.
Previously, the unspecified order of the methods just happened to make
the add operation work.

After this commit the tag list is refreshed when a new element is added
to it, in order to not rely on the order of methods anymore.

[1]: https://github.com/odoo/odoo/commit/a48a30f954afcb6ff3a59c4f32b05fd0c2cfcd2b

task-2811746