[FIX] website, web_editor: fix logo selector not displayed if text logo
Before this commit, the logo options (type, height, height scrolled)
were not displayed if the logo type was "Text".

After this commit, these options are moved to the navbar section so that
they are always displayed if the logo is "Text" or "Image". And if the
logo type is "None" only the logo type selector is displayed.

This commit fix it only in js because we are on a stable version, an
other commit will fix it in the snippets options XML in master.

task-2800680