[FIX] website: fix logo type selector not displayed if logo type is text
Before this commit, the logo type selector was not displayed in the
navbar options when the logo type was "Text".

After this commit, the logo type selector is moved in the navbar options
to be always displayed in any case. We also moved the height and height
scrolled logo options in the navbar options because otherwise those
options are not displayed if the logo type is "text".

task-2800680