[FIX] website: prevent animated navbar flicker
Dynamic headers change their layout as a function of the current scroll
position. However, in some edge cases, a change in the scroll position can also
be triggered by the layout change of the header. Generally speaking, this
happens when the decrease in the height of the header element is larger than
the scroll offset from the bottom of the page.

In such a case, more often than not, the navbar layout change directly triggers
another reverse navbar layout change, because of the decrease in scroll
position. Visually, the result is twofold: the user experiences a navbar
flicker and is generally also prevented from reaching the bottom scroll
position of the page.

However, the problem is easily resolved since there is already a system in
place that detaches the header and moves it to a fixed position. A padding
corresponding to the height of the header is then added to the main content,
preventing changes in the scroll height of the page. Making sure that this
padding always corresponds to the initial non-fixed state of the header
prevents the bug altogether.

opw-2812482