[REM] core: remove `_check_concurrency` method
`_check_concurrency` was used to check if 2 people was modify the
same record in the same time and it was able by putting in the context
`__last_update`: <value>.
It was unused seems lot of time and actually doesn't cover correctly the
case for the pending write.

Then remove it.