[IMP] website_event: include address in search of website event
Before, when we used to search on the event page, only the name of the event was
considered while searching.

After this  commit, the user will be able to search not only the Name of the
event but also the Address of the event.

We have achieved this by overriding the _search_fetch method of
website.searchable.mixin in event.event model where we are adapting the query
and adding a LEFT JOIN of res.partner and res.country for letting users search
the city and country respectively without using sudo().

Right now, we have only added the city and country of the address to the search
bar as those two fields are shown on the event page among all address fields.

TaskId-2791031