[WIP] base: convert WebP images to a format supported by wkhtmltopdf
Wkhtmltopdf is unable to render WebP images.
This commit intercepts the requests for the image data during PDF
reports rendering to replace them by a PNG.
Since we cannot use the WebP Pillow plugin for security concerns, the
conversion is achieved by an headless Chrome.

This WIP just copies the headless browser from odoo.tests.common.
If approved, a clean extract of the browser will be done instead.

task-2774352