[IMP] account,mail : Use job queue to send by mail invoices
This commit uses the mail queue when the user uses the "Send & Print"
functionality. The behavior before this commit is a synchronous send,
meaning that the app wait for EDI validation before sending the mail and
closing the wizard.

It creates a new model, mail_message_pending, that
represents a mail_message that is yet to be send and store all values
that will be used to send it.

We have used this approach because we wanted to put the message in the
chatter at the right time, i.e. when the mail is send (and not when the
message is created). Since the order of messages in the chatter is
linked to id (in js) it is the easiest way of doing it.

task-2764998