[FIX] crm: Lead 2 opportunities error when no salespersons In a multi-company environment, when we transform a lead (which has no salesperson) into an opportunity, we get an error.
To reproduce the issue:
1) Create a new company
2) Create a Lead for a customer with the Company set
3) Remove the Sales Team
4) Set the company on the Lead
5) Convert to an opportunity and you will see an error showing up

Solution:
A previous commit (1fad826743de9c8f916ef083e8de453212df6959) adapted the `_compute_company_id`
computation in which a case was forgotten (the case described here-above). The fix was to force
the `company_id` to the one of the `partner_id` if it has one.

opw-2805181