[ADD] missing_index: new module to control adding indexes to foreign keys
On a normal database, adding index to foreign keys is a good way to avoid systematic table scans in joins.
In odoo we do not create indexes on foreign key columns by default because of the size of those indexes.
Adding an index on all foreign keys, except 'create_uid' and 'write_uid' doesn't take more than 20% of database space.
Furthermore, keeping track of which index is used and removing the unused index is a good way to keep performance good and database size under control.

This module allows both to add indexes on foreign keys and to remove unused indexes (and re-add them if needed)

!!! this is the implem without the security