[FIX] web: hack for Bootstrap 5 for color
This commit is a hack for Bootstrap 5
In Bootstrap 4 these variables was used with map-merge()
In Bootstrap 5 these variables has set only if they don't exist before

PS: Another fix will be to reorder all bundle to follow Bootstrap
documentation and do another logic that wasn't use before.

Ref:
https://getbootstrap.com/docs/5.1/customize/sass/#colors

Note in this commit don't add to the `__manifest__.py` this file
as it must be included before `_variables.scss` of Bootstrap 5.
This will be done in another commit as it can be automated using RegEx.