[IMP]stock: create reverse move for return moves
As negative quantities are allowed into sale_orders (which would then
create a backorder when validated), changes in quantity aren't correctly
merged in the generated return picking.

In the case of a refund, instead of generating a negative move, instead
reverse the location_id/location_dest_id. Doing so would then try to
merge the move into a return picking.

Task-2659807