[IMP] sale: Only record down payments when confirmed
When creating down payments for sale orders the down payments
where shown even if the invoice was on a 'draft' status.
This might be confusing for users that cannot differentiate then
between draft payments and confirmed or paid down payments.
Now down payments in draft state are easily discernible.

Task - 2702407