[IMP] mail: try to better support multi emails when sending mails
Purpose of this commit is to get rid of custom formataddr used when sending
mails to partners. Two annoying use cases are managed here

  * multi emails ``email1@domain.com, email2@domain.com``
  * formatted emails ``"Full Name" <email@domain.com>``

Using formataddr leads to badly formatted emails. In this PR we fix that
computation by correctly setting multiple email_to, based on normalized
emails found in partner email field.

We do not use ``email_formatted`` as it is not really multi-enabled in
stable, preferring a small custom approach to be as tolerant as possible
with respect to user inputs.

Task-2612945 (Mail: Defensive email formatting)
Followup of odoo#54195