[FIX] mail: fix issues with formatting of email_from
As from should not contain multiple emails we now keep only the first found
one in email_from when sending emails. Indeed putting multi emails is possible
in Odoo fields, even if not officially supported. Let us be defensive and
avoid issues with the "from" of emails which should always be a single email.

Task-2612945 (Mail: Defensive email formatting)
Followup of odoo#54195