[IMP] point_of_sale: disable backend querying when all items loaded
In this commit, we improve the load-in-the-background
feature of partners/products.

If there are items that are not loaded yet, (either because the
background loading is still running or because the user did not
want to load any prodcts at all at the start of the session)
the option to fetch items from the database is enabled: every
time the user interacts with the search bar (pressing "Enter",
"Search database" or the "Search" icon at the search bar) the
backend is queried and new items (products/partners) are fetched
and added to the frontend. Before the fetching of products from
the backend was done using the "database" icon / "Load Customers"
button which are now dropped.

In this commit, the feature of querrying the database
is automatically disabled once all the partners or
products (available_in_pos) have been loaded in PoS.
Then, the querying of the backend stops.

task-2755384