[IMP] website: avoid SQL NOW() query to get create_date
When the ORM add missing default values on preparation of a create()
operation, it actually performs an SQL Query every time to get the NOW()
value.

See `def now(self):`.

There is a plan for the ORM to stop doing that and get that value in
python instead, but that's not yet done as this introduce some issues
that need to be investigated and solution found, especially in tests.

While waiting for that ORM change (if it happens), this will prevent
that SQL Query.